backend/requirements.txt
eventlet
python-socketio
pyjwt
mysql-connector-python